DB layer overhaul
I introduced an update method.
I wanted to use kwargs, both the SET & the WHERE data.
Pretty hard to do in 1 method.
I kinda refactored the DB layer: select() provides
WHERE data and is a prereq for update & delete.
But the result for select() can also be used (as iterable)

Updated other parts in the code to reflect to these changes.

Also fixed link.py - a previous commit where I got rid of
the call to GH broke it - I obviously still need it, to
add the hook.