Improve summary page
Moved "gadgets" to top, to avoid confusion that table values
have something to do with what's on the right (which is about current commit)

Removed progress chart: saves API call & opens up a spot for explanation of
the other things.

Clarify those other things.

Link to metrics for every commit.