Fix check for missing commits
That delete broke the check that followed.
That delete was intended to keep data in memory (and data that would
have to be looped over eventually) low once it is no longer needed,
but `commits` would be the same full array every time regardless,
so this made very little difference.